backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-jose[cryptography]>=3.3.0
httpx>=0.28.0
python-multipart>=0.0.18
anthropic>=0.42.0
pymupdf>=1.25.0
rapidfuzz>=3.11.0
structlog>=24.4.0
google-auth>=2.38.0
requests>=2.32.0
